admin-service/requirements.txt
fastapi
uvicorn
openai
python-dotenv
jinja2
websockets
pydantic
python-multipart
qdrant-client
langchain
langchain_community
langchain_text_splitters
langchain_qdrant
langchain-openai
langchain-docling
pypdf
boto3
beautifulsoup4
fastembed
docstring_parser
python-decouple
theflow
openai
anthropic
google-generativeai
aisuite[all]
tiktoken